requirements.txt
Django
South
python-dateutil
mock
